soulweb/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.5.0
pydantic==2.6.4
pydantic-core==2.16.3
email-validator==2.1.1
python-dotenv==1.0.1
firebase-admin==6.9.0
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.5.0